BT_HumanoidEntity redone
The humanoid entity's behavior tree was redone because it was too simple. Now the enemy starts wandering randomly around the area until it detects or sees the player, in which case it will start following them for 10 seconds, then remain still for approximately 2 seconds, before returning to the initial check. If it detects the player, it follows them; if not, it goes to a random location.