Several BPs fixes and Buoyoncy test
Fixed the life/damage system between the player and the Camel and Humanoid entities, corrected the HumanoidChase task, corrected the NavModifier of the poolroom, corrected the HHearSense of the Camel entity, corrected the tamalo of the Camel entity, added background music, and replaced the default sfx used by the Humanoid entity. A minor detail with the intensity of the lights in the Poolroom was fixed, and the level after death was changed to ParkingLot to facilitate video testing. Finally, the buoyancy simulation test was carried out in the poolroom's water tank using an inflatable ring. It is functional but needs a lot of improvement.