Replaced FPS default weapon and projectile
The default FPS gun was replaced by a personalized pistol with an adapted transform so the arms animations could be used.
The default projectile was changed with an actual bullet, replacing the sound effect and removing the bounciness property, the collision settings was also modified so the bullet doesn't get sticked in static meshes.